Please feel free to borrow some code if you like what you see
This repository also uses a MySQL80 Database. 
I added the DB script and a backup with some Data to use that you can restore.
I also use a CrystalDecisions library from NuGet to run a Crystal Report. I suggest you look for it and add it to this project.
You will see the reportDef in the /ReportDefs folder